+++ ntex/src/framed/mod.rs
@@ -7,10 +7,10 @@ mod time;
 mod write;
 
 pub use self::dispatcher::Dispatcher;
-pub use self::read::FramedReadTask;
+pub use self::read::ReadTask;
 pub use self::state::State;
 pub use self::time::Timer;
-pub use self::write::FramedWriteTask;
+pub use self::write::WriteTask;
 
 use crate::codec::{Decoder, Encoder};